mlx5: do not decompress mini CQEs
Mini CQEs are used by the NIC when the PCI bandwidth is limited, in this case,
the first CQE64 received contains almost everything and the number following
CQE8, following CQEs are empty and used by the software for decompression.

From this:

      0           1          2           6         7         8
  +-------+  +---------+ +-------+   +-------+ +-------+ +-------+
  | CQE64 |  |  CQE64  | | CQE64 |   | CQE64 | | CQE64 | | CQE64 |
  |-------|  |---------| |-------|   |-------| |-------| |-------|
  | ..... |  | cqe8[0] | |       | . |       | |       | | ..... |
  | ..... |  | cqe8[1] | |       | . |       | |       | | ..... |
  | ..... |  | ....... | |       | . |       | |       | | ..... |
  | ..... |  | cqe8[7] | |       |   |       | |       | | ..... |
  +-------+  +---------+ +-------+   +-------+ +-------+ +-------+

We get this after decompression:

      0          1     ...     8
  +-------+  +-------+     +-------+
  | CQE64 |  | CQE64 |     | CQE64 |
  |-------|  |-------|     |-------|
  | ..... |  | ..... |  .  | ..... |
  | ..... |  | ..... |  .  | ..... |
  | ..... |  | ..... |  .  | ..... |
  | ..... |  | ..... |     | ..... |
  +-------+  +-------+     +-------+

This decompression cost a lot in ressources knowing that the software only
needs the first CQE64 and the arrays of CQE8.

The change consists in getting the informations from the CQE8[n] and store it in
the previous CQE64 (the one used to inform about the compression).

Signed-off-by: Nelio Laranjeiro <nelio.laranjeiro@6wind.com>